Move MatchedSymbol, SemanticRule, and Action into parser_types.hpp.
